LPD-83280 Move revenue filter to separate component
