+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/resources/content/Language.properties
@@ -105,6 +105,7 @@ all-experiences=All Experiences
 all-fields-have-been-mapped-automatically-to-the-best-possible-match=All fields have been mapped automatically to the best possible match.
 all-interests=All Interests
 all-properties-from-this-workspace-have-access-to-this-data-source=All properties from this workspace have access to this data source.
+all-revenue-ranges=All Revenue Changes
 all-search-terms=All Search Terms
 all-users=All Users
 all-users-from-this-workspace-have-access-to-this-property=All users from this workspace have access to this property.
@@ -1199,6 +1200,10 @@ retention-period=Retention Period
 retrain=Retrain
 retraining-has-been-started=Retraining has been started.
 return-to-list=Return to List
+revenue=Revenue
+revenue-is-between=Revenue is Between
+revenue-is-greater-than=Revenue is Greater Than
+revenue-is-less-than=Revenue is Less Than
 review=Review
 review-&-run=Review & Run
 review-synced-data=Review Synced Data
